split out components file from views file
