+++ containers/tefca-viewer/src/app/query/views/CustomizeQueryView.tsx
@@ -2,7 +2,7 @@
 
 import React, { useMemo, useState, useEffect } from "react";
 import { Accordion, Button, Icon } from "@trussworks/react-uswds";
-import { AccordianSection } from "../../query/component-utils";
+import { AccordianSection } from "../component-utils";
 import { ValueSet } from "../../constants";
 import { AccordionItemProps } from "@trussworks/react-uswds/lib/components/Accordion/Accordion";
 import {
@@ -12,7 +12,7 @@ import {
 } from "@/app/database-service";
 import { UseCaseQueryResponse } from "@/app/query-service";
 import LoadingView from "./LoadingView";
-import { showRedirectConfirmation } from "./RedirectionToast";
+import { showRedirectConfirmation } from "../components/RedirectionToast";
 
 interface CustomizeQueryProps {
   useCaseQueryResponse: UseCaseQueryResponse;
@@ -59,7 +59,7 @@ const CustomizeQuery: React.FC<CustomizeQueryProps> = ({
   const handleSelectAllChange = (
     items: any[],
     setItems: React.Dispatch<React.SetStateAction<any[]>>,
-    checked: boolean,
+    checked: boolean
   ) => {
     const updatedItems = items.map((item) => ({ ...item, include: checked }));
     setItems(updatedItems);
@@ -68,7 +68,7 @@ const CustomizeQuery: React.FC<CustomizeQueryProps> = ({
   const handleIncludeAll = (
     setValueSet: React.Dispatch<React.SetStateAction<ValueSet>>,
     key: keyof ValueSet,
-    include: boolean,
+    include: boolean
   ) => {
     setValueSet((prevValueSet) => ({
       ...prevValueSet,
@@ -109,13 +109,13 @@ const CustomizeQuery: React.FC<CustomizeQueryProps> = ({
     const fetchQuery = async () => {
       const queryResults = await getSavedQueryByName(queryName);
       const labs = await mapQueryRowsToValueSetItems(
-        await filterQueryRows(queryResults, "labs"),
+        await filterQueryRows(queryResults, "labs")
       );
       const meds = await mapQueryRowsToValueSetItems(
-        await filterQueryRows(queryResults, "medications"),
+        await filterQueryRows(queryResults, "medications")
       );
       const conds = await mapQueryRowsToValueSetItems(
-        await filterQueryRows(queryResults, "conditions"),
+        await filterQueryRows(queryResults, "conditions")
       );
 
       // Only update if the fetch hasn't altered state yet
@@ -140,7 +140,7 @@ const CustomizeQuery: React.FC<CustomizeQueryProps> = ({
     const items = valueSetState[activeTab as keyof ValueSet];
     const selectedCount = items.filter((item) => item.include).length;
     const topCheckbox = document.getElementById(
-      "select-all",
+      "select-all"
     ) as HTMLInputElement;
     if (topCheckbox) {
       topCheckbox.indeterminate =
@@ -185,7 +185,7 @@ const CustomizeQuery: React.FC<CustomizeQueryProps> = ({
                           ...prevState,
                           [activeTab]: updatedItems,
                         })),
-                      selectedCount !== items.length,
+                      selectedCount !== items.length
                     );
                   }}
                 >
