+++ containers/tefca-viewer/src/app/query/views/LoadingView.tsx
